add controller logic so user can add, edit, delete own items only

+++ app/controllers/items_controller.rb
@@ -13,13 +13,11 @@ class ItemsController < ApplicationController
   end
 
   def new
-    result = current_user.restaurants.uniq.find{|restaurant| restaurant.id} == Restaurant.find(params[:restaurant_id])
-    if result == true
+    if current_user.restaurants.uniq.include?(Restaurant.find(params[:restaurant_id]))
       @item = Item.new
       @restaurant = Restaurant.find(params[:restaurant_id])
     else
-      @restaurant = Restaurant.find(params[:restaurant_id])
-      redirect_to restaurant_items_path(@restaurant)
+      redirect_to restaurants_path
     end
   end
 
@@ -45,13 +43,11 @@ class ItemsController < ApplicationController
   end
 
   def edit
-    result = current_user.restaurants.uniq.find{|restaurant| restaurant.id} == Restaurant.find(params[:restaurant_id])
-    if result
+    if current_user.restaurants.uniq.include?(Restaurant.find(params[:restaurant_id]))
       @item = Item.find(params[:id])
       @restaurant = Restaurant.find(params[:restaurant_id])
     else
-      @restaurant = Restaurant.find(params[:restaurant_id])
-      redirect_to restaurant_items_path(@restaurant)
+      redirect_to restaurants_path
     end
   end
 
@@ -61,16 +57,20 @@ class ItemsController < ApplicationController
     if @item.update(item_params)
         redirect_to restaurant_path(@item.restaurant_id)
     else
-      @restaurant = Restaurant.find(params[:item][:restaurant_id])
+      # @restaurant = Restaurant.find(params[:item][:restaurant_id])
       render :edit
     end
   end
 
   def destroy
-    item = Item.find(params[:id])
-    restaurant = Restaurant.find_by(id: item.restaurant_id)
-    item.destroy
-    redirect_to restaurant_path(restaurant)
+    if current_user.restaurants.uniq.include?(Restaurant.find(params[:restaurant_id]))
+      item = Item.find(params[:id])
+      restaurant = Restaurant.find_by(id: item.restaurant_id)
+      item.destroy
+      redirect_to restaurant_path(restaurant)
+    else
+      redirect_to restaurants_path
+    end
   end
 
   private
